Move storefront templatetags tests to separate file
